+++ apps/docs/content/components/index.ts
@@ -24,3 +24,4 @@ export * from "./radio-group";
 export * from "./switch";
 export * from "./popover";
 export * from "./tooltip";
+export * from "./tabs";