+++ packages/hooks/use-aria-field/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nextui-org/use-aria-field
 
+## 0.0.0-dev-v2-20230609000231
+
+### Patch Changes
+
+- Bordered colors changed to divider color
+- Updated dependencies
+  - @nextui-org/use-aria-slot-id@0.0.0-dev-v2-20230609000231
+  - @nextui-org/use-aria-label@0.0.0-dev-v2-20230609000231
+
 ## 0.0.0-dev-v2-20230607013542
 
 ### Patch Changes